Move RolesField into the authorization package
